Fix navigation bar styling and components
